requirements.txt
pdf2image==1.16.3
pytesseract==0.3.10
transformers==4.25.1
tensorflow==2.16.1  # Use this latest stable version
whoosh==2.7.4
Pillow==9.4.0
